requirements.txt
Django==5.2.1
django-crispy-forms==2.4
Pillow==11.2.1
django-ckeditor==6.7.2
django-environ==0.12.0
mysqlclient==2.2.7
djangorestframework==3.16.0